backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.19
pymysql==1.1.0
minio==7.1.15
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyyaml==6.0.1
python-docx==0.8.11
jinja2==3.1.2
weasyprint==59.0
pandas==2.0.3
matplotlib==3.7.2
reportlab==4.0.4
PyPDF2==3.0.1
Pillow==10.0.0
cryptography==41.0.2
requests==2.31.0
